Removed jolt cooking dll
