--- Code/EnginePlugins/JoltCooking/JoltCookingPCH.cpp
@@ -1,1 +1,0 @@
-#include <JoltCooking/JoltCookingPCH.h>
