--- Code/EnginePlugins/JoltCooking/JoltCookingPCH.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <Foundation/Basics.h>
-#include <Foundation/Logging/Log.h>
-
-#include <Jolt/Jolt.h>
