+++ Code/EnginePlugins/JoltPlugin/Components/Implementation/JoltRagdollComponent.cpp
@@ -92,8 +92,6 @@ EZ_BEGIN_COMPONENT_TYPE(ezJoltRagdollComponent, 5, ezComponentMode::Dynamic)
 EZ_END_ABSTRACT_COMPONENT_TYPE;
 // clang-format on
 
-EZ_DEFINE_AS_POD_TYPE(JPH::Vec3);
-
 //////////////////////////////////////////////////////////////////////////
 
 ezJoltRagdollComponentManager::ezJoltRagdollComponentManager(ezWorld* pWorld)
